fix: respect key case, elide shift only when case matches
This fixes a regression in the eventToHotkeyString function which was
caused by trying to normalise the event.key toUpperCase. The regression
was because on macos the "Meta+Shift" plane is always lowercase, which
is contrary to other operating systems. This is a regression because
that was expected behavior as part of this package, and the fix in
1.4.3 broke this expectation while trying to fix another bug.

This change respects the `event.key` spacing, and elides the shift key
only when the event.key matches case, returning `Meta+Shift+p` as a
viable shortcut.